6. ResponseEntity. Статусы ответов. [#505111]

+++ src/main/java/ru/job4j/media/controller/UserController.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import ru.job4j.media.model.User;
 import ru.job4j.media.service.UserService;
 
@@ -15,7 +16,15 @@ public class UserController {
 
     @PostMapping
     public ResponseEntity<User> save(@RequestBody User user) {
-        return ResponseEntity.ok(userService.save(user));
+        userService.save(user);
+        var uri = ServletUriComponentsBuilder
+                .fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(user.getId())
+                .toUri();
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .location(uri)
+                .body(user);
     }
 
     @GetMapping("/{userId}")
@@ -26,14 +35,19 @@ public class UserController {
     }
 
     @PutMapping
-    @ResponseStatus(HttpStatus.OK)
-    public void update(@RequestBody User user) {
-        userService.save(user);
+    public ResponseEntity<Void> update(@RequestBody User user) {
+        if (userService.save(user) != null) {
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{userId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable int userId) {
+    public ResponseEntity<Void> delete(@PathVariable int userId) {
+        if (userService.findById(userId).isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         userService.delete(userId);
+        return ResponseEntity.noContent().build();
     }
 }
